fix: замена stateful виджета с настройками игры на change notifier для более удобного доступа из любой части поддерева виджетов
