+++ frontend/lib/views/setting_view/components/settings_rows_section.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:frontend/common/shared_functions.dart';
 import 'package:frontend/exports.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:frontend/views/setting_view/providers/game_settings_provider.dart';
+import 'package:provider/provider.dart';
 
 class SettingsRowsSection extends StatelessWidget {
   const SettingsRowsSection({
@@ -10,7 +12,6 @@ class SettingsRowsSection extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final settingsState = gameSettingsViewStateKey.currentState;
     final l10n = AppLocalizations.of(context)!;
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
@@ -27,28 +28,29 @@ class SettingsRowsSection extends StatelessWidget {
             ),
           ),
         ),
-        Column(
-          children: [
-            SettingsRow(
-              chose: settingsState?.isMoveBack,
-              text: l10n.undoMoves,
-              modalHeader: l10n.undoMovesDescription,
-              onChanged: settingsState?.setIsMoveBack,
-            ),
-            SettingsRow(
-              chose: settingsState?.isThreats,
-              text: l10n.threats,
-              modalHeader: l10n.threatsDescription,
-              onChanged: settingsState?.setIsThreats,
-            ),
-            SettingsRow(
-              chose: settingsState?.isHints,
-              text: l10n.hints,
-              modalHeader: l10n.hintsDescription,
-              onChanged: settingsState?.setIsHints,
-            )
-          ],
-        ),
+        Consumer<GameSettingsProvider>(
+            builder: (_, provider, __) => Column(
+                  children: [
+                    SettingsRow(
+                      chose: provider.isMoveBack,
+                      text: l10n.undoMoves,
+                      modalHeader: l10n.undoMovesDescription,
+                      onChanged: provider.setIsMoveBack,
+                    ),
+                    SettingsRow(
+                      chose: provider.isThreats,
+                      text: l10n.threats,
+                      modalHeader: l10n.threatsDescription,
+                      onChanged: provider.setIsThreats,
+                    ),
+                    SettingsRow(
+                      chose: provider.isHints,
+                      text: l10n.hints,
+                      modalHeader: l10n.hintsDescription,
+                      onChanged: provider.setIsHints,
+                    )
+                  ],
+                )),
       ],
     );
   }
